document cli pillar and alter cmd.run states
The readme didn't have an example pillar for CLI and therefore it caused confusion when the cli tool was an empty executable file.

This caused confusion for new users.

In addtion new versions of Salt Stack replaced cmd.run's `user` argument with `runas`.

As a result many states in this formula drifted and no longer work.

	modified:   README.rst
	modified:   wordpress/config.sls
	modified:   wordpress/init.sls
	modified:   wordpress/plugin.sls